Season Policy - ep 108

+++ modules/Cyaxaress/Course/Http/Controllers/SeasonController.php
@@ -4,7 +4,9 @@ namespace Cyaxaress\Course\Http\Controllers;
 
 use Cyaxaress\Common\Responses\AjaxResponses;
 use Cyaxaress\Course\Http\Requests\SeasonRequest;
+use Cyaxaress\Course\Models\Course;
 use Cyaxaress\Course\Models\Season;
+use Cyaxaress\Course\Repositories\CourseRepo;
 use Cyaxaress\Course\Repositories\SeasonRepo;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -17,8 +19,9 @@ class SeasonController extends Controller
         $this->seasonRepo = $seasonRepo;
     }
 
-    public function store($course, SeasonRequest $request )
+    public function store($course, SeasonRequest $request, CourseRepo $courseRepo)
     {
+        $this->authorize('createSeason', $courseRepo->findByid($course));
         $this->seasonRepo->store($course, $request);
         newFeedback();
         return back();
@@ -27,11 +30,13 @@ class SeasonController extends Controller
     public function edit($id)
     {
         $season = $this->seasonRepo->findByid($id);
+        $this->authorize('edit', $season);
         return view('Courses::seasons.edit', compact('season'));
     }
 
     public function update($id, SeasonRequest $request)
     {
+        $this->authorize('edit', $this->seasonRepo->findByid($id));
         $this->seasonRepo->update($id, $request);
         newFeedback();
         return back();
@@ -39,7 +44,7 @@ class SeasonController extends Controller
 
     public function accept($id)
     {
-//        $this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateConfirmationStatus($id, Season::CONFIRMATION_STATUS_ACCEPTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -49,7 +54,7 @@ class SeasonController extends Controller
 
     public function reject($id)
     {
-//        $this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateConfirmationStatus($id, Season::CONFIRMATION_STATUS_REJECTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -59,7 +64,7 @@ class SeasonController extends Controller
 
     public function lock($id)
     {
-//        $this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateStatus($id, Season::STATUS_LOCKED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -68,7 +73,7 @@ class SeasonController extends Controller
     }
     public function unlock($id)
     {
-//        $this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateStatus($id, Season::STATUS_OPENED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -78,6 +83,7 @@ class SeasonController extends Controller
     public function destroy($id)
     {
         $season = $this->seasonRepo->findByid($id);
+        $this->authorize('delete', $season);
 
         $season->delete();
